# backend-go/docs/swagger.yaml
basePath: /api
definitions:
  features_clubs_presentation.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  presentation.AddMemberRequest:
    properties:
      userSlug:
        type: string
    required:
    - userSlug
    type: object
...
  presentation.ClubMembershipResponse:
    properties:
      clubId:
        type: integer
      clubName:
        type: string
      clubSlug:
        type: string
      createdAt:
        type: string
      endDate:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      nextBillingDate:
        type: string
      paymentStatus:
        type: string
      startDate:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userEmail:
        type: string
      userId:
        type: integer
      userName:
        type: string
      userSlug:
        type: string
    type: object
  presentation.ClubResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        type: integer
      memberCount:
        type: integer
      monthlyFeeCents:
        type: integer
      monthlyFeeEuros:
        type: number
      name:
        type: string
      ownerId:
        type: integer
      ownerName:
        type: string
      ownerSlug:
        type: string
      slug:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  presentation.CreateClubRequest:
    properties:
      description:
        type: string
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        minimum: 1
        type: integer
      monthlyFeeCents:
        minimum: 0
        type: integer
      name:
        type: string
      ownerSlug:
        type: string
    required:
    - maxMembers
    - name
    type: object
...
  presentation.RenewMembershipRequest:
    properties:
      customerId:
        type: string
    required:
    - customerId
    type: object
...
  presentation.UpdateBillingDateRequest:
    properties:
      nextBillingDate:
        type: string
    required:
    - nextBillingDate
    type: object
...
  presentation.UpdateClubRequest:
    properties:
      description:
        type: string
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        minimum: 1
        type: integer
      monthlyFeeCents:
        minimum: 0
        type: integer
      name:
        type: string
      ownerSlug:
        type: string
      status:
        type: string
    required:
    - maxMembers
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API REST para gestión de instalaciones deportivas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: PoliManage API
  version: "1.0"
paths:
...
  /clubs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.ClubResponse'
            type: array
      summary: Lista todos los clubs
      tags:
      - Clubs
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del club
        in: body
        name: club
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Crea un nuevo club
      tags:
      - Clubs
  /clubs/{slug}:
    delete:
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Elimina un club
      tags:
      - Clubs
    get:
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Obtiene un club por slug
      tags:
      - Clubs
    put:
      consumes:
      - application/json
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      - description: Datos del club
        in: body
        name: club
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateClubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Actualiza un club existente
      tags:
      - Clubs
  /clubs/{slug}/members:
    get:
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.ClubMembershipResponse'
            type: array
      summary: Obtiene los miembros de un club
      tags:
      - Clubs
    post:
      consumes:
      - application/json
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      - description: Datos del miembro
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/presentation.AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.ClubMembershipResponse'
      summary: Añade un miembro a un club
      tags:
      - Clubs
  /clubs/memberships/{id}:
    delete:
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Elimina un miembro de un club
      tags:
      - Clubs
  /clubs/memberships/{id}/billing-date:
    put:
      consumes:
      - application/json
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nueva fecha
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateBillingDateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_clubs_presentation.MessageResponse'
      summary: Actualiza la fecha de próximo cobro
      tags:
      - Clubs
  /clubs/memberships/{id}/cancel:
    post:
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_clubs_presentation.MessageResponse'
      summary: Cancela permanentemente una membresía
      tags:
      - Clubs
  /clubs/memberships/{id}/renew:
    post:
      consumes:
      - application/json
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del pago
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presentation.RenewMembershipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_clubs_presentation.MessageResponse'
      summary: Renueva una membresía procesando el pago
      tags:
      - Clubs
  /clubs/memberships/{id}/resume:
    post:
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_clubs_presentation.MessageResponse'
      summary: Reanuda una membresía suspendida
      tags:
      - Clubs
  /clubs/memberships/{id}/suspend:
    post:
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_clubs_presentation.MessageResponse'
      summary: Suspende una membresía
      tags:
      - Clubs
schemes:
- http
swagger: "2.0"